Revert "Merge branch 'back-feat-admin' into 'back'"
This reverts merge request !77

+++ backend/controller/site-controller.js
@@ -2,7 +2,7 @@ import is from "@sindresorhus/is";
 import { siteService } from "../services";
 import { asyncHandler } from "../utils";
 
-import { BadRequestError } from "../errors";
+import { BadRequestError, ForbiddenError } from "../errors";
 
 const siteController = {
   addsite: asyncHandler(async (req, res, next) => {
@@ -11,7 +11,6 @@ const siteController = {
         "headers의 Content-Type을 application/json으로 설정해주세요"
       );
     }
-    console.log(req.body);
     const { userId, name, domain, theme, font, colorset, blocks } = req.body;
     const newSite = await siteService.addSite({
       userId,
@@ -26,12 +25,22 @@ const siteController = {
   }),
 
   getSiteInfo: asyncHandler(async (req, res) => {
+    if (is.emptyObject(req.body)) {
+      throw new BadRequestError(
+        "headers의 Content-Type을 application/json으로 설정해주세요"
+      );
+    }
     const siteId = req.params.siteId;
     const site = await siteService.getSiteInfo(siteId);
     return res.ok(200, site);
   }),
 
   getUserSites: asyncHandler(async (req, res) => {
+    if (is.emptyObject(req.body)) {
+      throw new BadRequestError(
+        "headers의 Content-Type을 application/json으로 설정해주세요"
+      );
+    }
     const userId = req.params.userId;
     const sites = await siteService.getSites(userId);
     return res.ok(200, sites);
@@ -64,6 +73,11 @@ const siteController = {
   }),
 
   deleteSite: asyncHandler(async (req, res) => {
+    if (is.emptyObject(req.body)) {
+      throw new BadRequestError(
+        "headers의 Content-Type을 application/json으로 설정해주세요"
+      );
+    }
     const siteId = req.params.siteId;
     const deletedSite = await siteService.deleteSiteBySiteId(siteId);
     return res.ok(200, deletedSite);
